DB classes

+++ includes/DB/Category.php
@@ -36,7 +36,7 @@ class Category extends Database{
 
     }
 
-    public function all(){
+    public function allCategories(){
         $categories = array();
         $sql = "SELECT * FROM categories";
         $result = $this->connection->query($sql) or die($this->connection->error);
@@ -47,15 +47,58 @@ class Category extends Database{
         return $categories;
     }
 
+    public function allItems($categoryId){
+        $items = array();
+        $sql = "SELECT * FROM items WHERE categoryId={$categoryId}";
+        $result = $this->connection->query($sql) or die($this->connection->error);
+        while($row = $result->fetch_assoc()) {
+            $items[]= array ($row["id"],$row['imgName'],$row['itemName'],$categoryId);
+
+        }
+        return $items;
+    }
+
     public function addItem($img,$itemName,$categoryId){
-        $sql = "INSERT INTO items (imgName,itemName,categoryId) VALUES ('{$img}','{$itemName}','{$categoryId}')";
-        if (mysqli_query($this->connection, $sql)) {
-            echo "New record created successfully";
+        if(!$this->itemExists($itemName)){
+            $sql = "INSERT INTO items (imgName,itemName,categoryId) VALUES ('{$img}','{$itemName}','{$categoryId}')";
+            if (mysqli_query($this->connection, $sql)) {
+                echo "New record created successfully";
+                return true;
+            } else {
+                echo "Error: " . $sql . "<br>" . mysqli_error($this->connection);
+                return false;
+            }
+        }
+    }
+
+    public function getItem($id){
+        $sql = "SELECT * FROM items WHERE id={$id}";
+        $result = $this->connection->query($sql) or die($this->connection->error);
+        $row = $result->fetch_assoc();
+        return $row;
+    }
+
+    public function itemExists($itemName){
+        $sql = "SELECT * FROM items WHERE itemName='{$itemName}'";
+        $result = $this->connection->query($sql) or die($this->connection->error);
+
+        if ($result->num_rows > 0) {
             return true;
         } else {
-            echo "Error: " . $sql . "<br>" . mysqli_error($this->connection);
             return false;
         }
+
+    }
+
+    public function getAllItems(){
+        $items = array();
+        $sql = "SELECT * FROM items";
+        $result = $this->connection->query($sql) or die($this->connection->error);
+        while($row = $result->fetch_assoc()) {
+            $items[]= array ($row["id"],$row['imgName'],$row['itemName'],$row['categoryId']);
+
+        }
+        return $items;
     }
 
 }
